requirements.txt
numpy==1.26.4
psutil==5.9.8
GPUtil==1.4.0
chex==0.1.86
gymnasium==0.29.1
mujoco==3.1.6
dm-control==1.0.20
torch==2.3.1
pandas==2.2.2
pillow==10.4.0
ipdb==0.13.13
wandb==0.17.3
moviepy==1.0.3
imageio==2.34.2
opencv-python==4.10.0.84
ruamel.yaml==0.18.6
rich==13.7.1
cloudpickle==3.0.0
ml-dtypes==0.3.2
tensorflow==2.16.2
tensorflow-probability==0.24.0
dm-sonnet==2.0.2
optax==0.2.2
plotly==5.22.0
msgpack==1.0.8
pyzmq==26.0.3
colored==2.2.4
matplotlib==3.9.0
stable-baselines3[extra]>=2.3.0
